multimodal/requirements.txt
PyMuPDF
Pillow
python-dotenv
fastapi
uvicorn
python-multipart
groq
